requirements.txt
flask>=3.0
flask-sock>=0.7
anthropic>=0.40
python-dotenv>=1.0
gunicorn>=21.0
